requirements.txt
Click==7.0
defusedxml==0.6.0
Genshi==0.7.3
lxml==4.5.0
odfpy==1.4.1
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyqt5-tools==5.13.2.1.6rc1
python-dotenv==0.10.5
